+++ src/YandexDirect/Api/V5/Contract/CampaignUpdateItem.php
@@ -30,6 +30,8 @@ class CampaignUpdateItem extends CampaignBase
 
     protected $DynamicTextCampaign = null;
 
+    protected $TimeTargeting = null;
+
     /**
      * Creates a new instance of CampaignUpdateItem.
      *
@@ -289,6 +291,29 @@ class CampaignUpdateItem extends CampaignBase
     public function setDynamicTextCampaign(DynamicTextCampaignUpdateItem $value = null)
     {
         $this->DynamicTextCampaign = $value;
+
+        return $this;
+    }
+
+    /**
+     * Gets TimeTargeting.
+     *
+     * @return TimeTargeting|null
+     */
+    public function getTimeTargeting()
+    {
+        return $this->TimeTargeting;
+    }
+
+    /**
+     * Sets TimeTargeting.
+     *
+     * @param TimeTargeting|null $value
+     * @return $this
+     */
+    public function setTimeTargeting(TimeTargeting $value = null)
+    {
+        $this->TimeTargeting = $value;
 
         return $this;
     }